# Speed up `Line2D._edit_get_rect()`