This method is called constantly on viewport movement and when `Line2D` node is selected.
Current implementation is not bottleneck for editor, but objectively not as effective as it could be even without doing over-optimization.
This PR replaces excessive `expand_to()` calls in for-loop with finding minimum and maximum positions among all points and then creating `Rect2D`. 

For one `Line2D` node with 116 points, single `Line2D._edit_get_rect()` call.

| Editor build | Before, average time, ns | After, average time, ns | Speed gain |
| --- | --- | --- | --- |
| Dev build | 9048 | 4756  | ~x1.9 faster |
| Production build | 2405 | 1150 | ~x2.1 faster |