@Ivorforce 2025-01-03
> Results are for dev build, so difference in optimized build may be not so big, but still should be.

I definitely recommend running the test on a non-dev build, because non-optimized builds are rarely representative of optimized ones. 

`expand_to ` is implemented in `rect.h` and thus fully visible to the compiler. Since it's dealing with floats, it's unlikely the call will be perfectly optimized - but it's good practice to check anyway :)