@arkology 2025-01-03
> I definitely recommend running the test on a non-dev build, because non-optimized builds are rarely representative of optimized ones.

Sure! Added results to OP. 

> `expand_to ` is implemented in `rect.h` and thus fully visible to the compiler. _Theoretically_, that means it could be optimized away - it probably isn't because it's dealing with floats - but it's good practice to check anyway :)

`min()`/`max()` are also implemented in header so I expected to get comparable results (and the are comparable).